+++ packages/video_player/video_player_avfoundation/test/avfoundation_video_player_test.dart
@@ -306,7 +306,7 @@ void main() {
                     mockChannel,
                     const StandardMethodCodec()
                         .encodeSuccessEnvelope(<String, dynamic>{
-                      'event': 'playingUpdate',
+                      'event': 'isPlayingStateUpdate',
                       'isPlaying': true,
                     }),
                     (ByteData? data) {});
@@ -317,7 +317,7 @@ void main() {
                     mockChannel,
                     const StandardMethodCodec()
                         .encodeSuccessEnvelope(<String, dynamic>{
-                      'event': 'playingUpdate',
+                      'event': 'isPlayingStateUpdate',
                       'isPlaying': false,
                     }),
                     (ByteData? data) {});
@@ -354,11 +354,11 @@ void main() {
             VideoEvent(eventType: VideoEventType.bufferingStart),
             VideoEvent(eventType: VideoEventType.bufferingEnd),
             VideoEvent(
-              eventType: VideoEventType.playingUpdate,
+              eventType: VideoEventType.isPlayingStateUpdate,
               isPlaying: true,
             ),
             VideoEvent(
-              eventType: VideoEventType.playingUpdate,
+              eventType: VideoEventType.isPlayingStateUpdate,
               isPlaying: false,
             ),
           ]));