+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.5.2
 
-* Synchronize `VideoPlayerValue.isPlaying` with underlying video player.
+* Synchronizes `VideoPlayerValue.isPlaying` with underlying video player.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.5.1