+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 6.0.2
 
-* Add the `VideoEventType.playingUpdate` event to track changes in play / pause status with the
-underlying video player.
+* Add the `VideoEventType.isPlayingStateUpdate` event to track changes in play / pause state with
+the underlying video player.
 * Updates minimum Flutter version to 3.0.
 
 ## 6.0.1