+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.14
 
-* Synchronize `VideoPlayerValue.isPlaying` with `VideoElement`.
+* Synchronizes `VideoPlayerValue.isPlaying` with `VideoElement`.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.0.13